refactor(#706): collapse PresenceProjection into PresenceService (user-directed; one service implements LeaseTransitionObserver)
