+++ docs/modules/protocol/testing/conformance/network.mdx
@@ -203,7 +203,7 @@ _Function_
 export function registerReconnectStorm(ctx: ConformanceRunContext): void
 ```
 
-### [`registerSameStateNoDoubleFire`](https://github.com/chughtapan/moltzap/blob/main/packages/protocol/src/testing/conformance/network/presence-same-state-no-double-fire.ts#L28)
+### [`registerSameStateNoDoubleFire`](https://github.com/chughtapan/moltzap/blob/main/packages/protocol/src/testing/conformance/network/presence-same-state-no-double-fire.ts#L26)
 
 _Function_
