+++ packages/server/src/network/handlers/presence.handlers.ts
@@ -4,19 +4,16 @@ import { PresenceSubscribe, NotInContactsError } from "@moltzap/protocol";
 import type { AgentId as ServerAgentId } from "../../app/types.js";
 import { Effect } from "effect";
 import { ConnectionTag, DbTag, PresenceServiceTag } from "../../app/layers.js";
-import { PresenceProjectionTag } from "../services/presence-projection.js";
 import { visibleAgentIds } from "../../identity/services/agent-visibility.js";
 
 /**
  * Presence RPC handlers.
  *
- * **Architect plan #706 v7 (codex r6 P2 #2) — narrowed surface.** The
- * `PresenceUpdate` RPC handler was deleted along with the `PresenceUpdate`
- * descriptor (impl-staff lands the protocol-package deletion + barrel
- * cleanup as part of the §8 cutover). The surviving `presence/subscribe`
- * handler reads status snapshots from `PresenceProjection.statusMany`
- * (v3 codex r2 P2 #1) instead of the deleted `PresenceService.getMany`;
- * subscriber-registry mutation stays on `PresenceService.subscribe`.
+ * `presence/subscribe` registers fan-out interest via
+ * `PresenceService.subscribe` and reads the current status snapshot via
+ * `PresenceService.statusMany`. Presence is server-derived from
+ * `LeaseRegistry` lifecycle + WS connect/disconnect; there is no
+ * client-driven `presence/update`.
  */
 export const presenceHandlers: RpcMethodRegistry = [
   defineTaskMethod(PresenceSubscribe, {
@@ -26,7 +23,6 @@ export const presenceHandlers: RpcMethodRegistry = [
     handler: (params, ctx) =>
       Effect.gen(function* () {
         const presenceService = yield* PresenceServiceTag;
-        const presenceProjection = yield* PresenceProjectionTag;
         const db = yield* DbTag;
         const connection = yield* ConnectionTag;
         const requested = params.agentIds as ServerAgentId[];
@@ -44,12 +40,10 @@ export const presenceHandlers: RpcMethodRegistry = [
           );
         }
         presenceService.subscribe(connection.id, visibleIds);
-        // v7 (codex r6 P2 #2): status snapshot from the projection,
-        // not the deleted PresenceService.getMany. The projection's
-        // result is a ReadonlyArray; the wire schema expects a
-        // mutable Array, so we spread to a fresh array (one-frame
-        // copy; safe — the snapshot is point-in-time).
-        const projected = yield* presenceProjection.statusMany(visibleIds);
+        // The status snapshot is a ReadonlyArray; the wire schema
+        // expects a mutable Array, so map to a fresh array (point-in-time
+        // copy).
+        const projected = yield* presenceService.statusMany(visibleIds);
         const statuses = projected.map((entry) => ({
           agentId: entry.agentId,
           status: entry.status,
