+++ packages/server/src/network/services/presence.service.test.ts
@@ -1,14 +1,13 @@
-/* eslint-disable agent-code-guard/no-example-only-tests, agent-code-guard/no-hardcoded-assertion-literals, max-lines-per-function, sonarjs/max-lines-per-function, max-nested-callbacks -- regression-only suite per architect plan #706 §8: each case names a contractual transition (online/working/offline, dedup, fast-reconnect race). PresenceStatus literals are contractual wire values pinned by the plan; making them imports would lose the regression intent. Vitest `describe`/`it` + Effect.gen naturally nest 4 deep. */
+/* eslint-disable agent-code-guard/no-example-only-tests, agent-code-guard/no-hardcoded-assertion-literals, max-lines-per-function, sonarjs/max-lines-per-function, max-nested-callbacks -- regression-only suite: each case names a contractual transition (online/working/offline, dedup, fast-reconnect race). PresenceStatus literals are contractual wire values; making them imports would lose the regression intent. Vitest `describe`/`it` + Effect.gen naturally nest 4 deep. */
 
 /**
- * Unit tests for the `PresenceProjection` module (architect plan #706,
- * impl-staff sub-issue #712).
+ * Unit tests for the `PresenceService` status engine.
  *
  * Covers state derivation (online / working / offline), subscriber
- * fan-out, dedup discipline, and the named regression cases from §8 of
- * the v6+ plan: redundant-connect-while-working,
- * fast-reconnect-during-abandon, concurrent-grant-during-fast-reconnect,
- * and lease callbacks on absent / stale-connection entries.
+ * fan-out, dedup discipline, and the named regression cases:
+ * redundant-connect-while-working, fast-reconnect-during-abandon,
+ * concurrent-grant-during-fast-reconnect, and lease callbacks on
+ * absent / stale-connection entries.
  */
 
 import { describe, expect } from "vitest";
@@ -25,10 +24,6 @@ import {
   type MoltZapConnection,
 } from "../../transport/connection.js";
 import { PresenceService } from "./presence.service.js";
-import {
-  makePresenceProjection,
-  type PresenceProjection,
-} from "./presence-projection.js";
 
 const it = effectIt.live;
 
@@ -42,13 +37,11 @@ const LEASE_Y = "lease-y" as LeaseId;
 
 interface Harness {
   readonly connections: ConnectionManager;
-  readonly subscribers: PresenceService;
   readonly sent: string[];
 }
 
 function makeHarness(): Harness {
   const connections = new ConnectionManager();
-  const subscribers = new PresenceService();
   const sent: string[] = [];
   const subscriberConn: MoltZapConnection = {
     id: SUBSCRIBER,
@@ -64,7 +57,7 @@ function makeHarness(): Harness {
     originator: {} as MoltZapConnection["originator"],
   };
   connections.add(subscriberConn);
-  return { connections, subscribers, sent };
+  return { connections, sent };
 }
 
 interface DecodedFrame {
@@ -88,42 +81,48 @@ function decodeFrames(
 }
 
 /**
- * Drain the microtask queue so the projection's fire-and-forget
- * `Effect.runFork` writes to the recording connection have landed
- * before assertions read `sent`. The `Effect.yieldNow` calls
- * relinquish the fiber so any forked writes that resolved
- * synchronously have a chance to interleave.
+ * Drain the microtask queue so the fire-and-forget `Effect.runFork`
+ * writes to the recording connection have landed before assertions
+ * read `sent`. The `Effect.yieldNow` calls relinquish the fiber so any
+ * forked writes that resolved synchronously have a chance to
+ * interleave.
  */
 const drainMicrotasks: Effect.Effect<void> = Effect.gen(function* () {
   yield* Effect.yieldNow();
   yield* Effect.yieldNow();
 });
 
+/**
+ * Build the service, register the given subscriber-watch sets, then run
+ * `body`. The subscriber registry and the status engine are one
+ * `PresenceService` instance, so subscriptions and transitions exercise
+ * the same object.
+ */
 function withHarness<A>(
   h: Harness,
-  body: (projection: PresenceProjection) => Effect.Effect<A>,
+  watch: ReadonlyArray<readonly [ConnectionId, ReadonlyArray<AgentId>]>,
+  body: (presence: PresenceService) => Effect.Effect<A>,
 ): Effect.Effect<A> {
   return Effect.gen(function* () {
-    const projection = yield* makePresenceProjection({
-      connections: h.connections,
-      subscribers: h.subscribers,
-    });
-    return yield* body(projection);
+    const presence = yield* PresenceService.make(h.connections);
+    for (const [connId, agentIds] of watch) {
+      presence.subscribe(connId, agentIds);
+    }
+    return yield* body(presence);
   });
 }
 
-describe("PresenceProjection — state derivation", () => {
+describe("PresenceService — state derivation", () => {
   it("WS connect emits `online`", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(decodeFrames(h.sent)).toEqual([
           { agentId: AGENT_A, status: "online" },
         ]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("online");
       }),
     );
@@ -131,18 +130,17 @@ describe("PresenceProjection — state derivation", () => {
 
   it("lease GRANTED while connected emits `working`", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(decodeFrames(h.sent)).toEqual([
           { agentId: AGENT_A, status: "working" },
         ]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("working");
       }),
     );
@@ -150,19 +148,18 @@ describe("PresenceProjection — state derivation", () => {
 
   it("lease CONSUMED while connected emits `online`", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(decodeFrames(h.sent)).toEqual([
           { agentId: AGENT_A, status: "online" },
         ]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("online");
       }),
     );
@@ -170,18 +167,17 @@ describe("PresenceProjection — state derivation", () => {
 
   it("WS close emits `offline`", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onAgentDisconnect(AGENT_A, CONN_1);
+        yield* presence.onAgentDisconnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(decodeFrames(h.sent)).toEqual([
           { agentId: AGENT_A, status: "offline" },
         ]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("offline");
       }),
     );
@@ -189,11 +185,11 @@ describe("PresenceProjection — state derivation", () => {
 
   it("`statusMany` returns per-agent snapshots; unknown agents resolve to `offline`", () => {
     const h = makeHarness();
-    return withHarness(h, (projection) =>
+    return withHarness(h, [], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
-        const result = yield* projection.statusMany([AGENT_A, AGENT_B]);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        const result = yield* presence.statusMany([AGENT_A, AGENT_B]);
         expect(result).toEqual([
           { agentId: AGENT_A, status: "working" },
           { agentId: AGENT_B, status: "offline" },
@@ -203,20 +199,19 @@ describe("PresenceProjection — state derivation", () => {
   });
 });
 
-describe("PresenceProjection — dedup discipline", () => {
+describe("PresenceService — dedup discipline", () => {
   it("second GRANTED lease while already `working` emits NO `presence/changed`", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onLeaseActiveBegin(LEASE_Y, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_Y, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("working");
       }),
     );
@@ -224,18 +219,17 @@ describe("PresenceProjection — dedup discipline", () => {
 
   it("only the LAST CONSUMED lease drops the agent back to `online`", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_Y, AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_Y, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        yield* projection.onLeaseActiveEnd(LEASE_Y, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveEnd(LEASE_Y, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(decodeFrames(h.sent)).toEqual([
           { agentId: AGENT_A, status: "online" },
@@ -245,20 +239,19 @@ describe("PresenceProjection — dedup discipline", () => {
   });
 });
 
-describe("PresenceProjection — regression cases (architect plan §8)", () => {
+describe("PresenceService — regression cases", () => {
   it("redundant-connect-while-working: second onAgentConnect with the same connId is a no-op", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("working");
       }),
     );
@@ -266,20 +259,19 @@ describe("PresenceProjection — regression cases (architect plan §8)", () => {
 
   it("fast-reconnect-during-abandon: stale lease callbacks audit without re-emitting", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
-        yield* projection.onAgentDisconnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onAgentDisconnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
-        yield* projection.onAgentConnect(AGENT_A, CONN_2);
+        yield* presence.onAgentConnect(AGENT_A, CONN_2);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("online");
       }),
     );
@@ -287,18 +279,17 @@ describe("PresenceProjection — regression cases (architect plan §8)", () => {
 
   it("concurrent-grant-during-fast-reconnect: stale onLeaseActiveBegin audits without re-emitting", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onAgentDisconnect(AGENT_A, CONN_1);
-        yield* projection.onAgentConnect(AGENT_A, CONN_2);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentDisconnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_2);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onLeaseActiveBegin(LEASE_Y, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_Y, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("online");
       }),
     );
@@ -306,44 +297,41 @@ describe("PresenceProjection — regression cases (architect plan §8)", () => {
 
   it("lease begin / end on an absent entry audits LeaseBeginAfterDisconnect / LeaseEndAfterDisconnect", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("offline");
       }),
     );
   });
 
-  it("stale-disconnect (entry.connId mismatch) is a silent no-op", () => {
+  it("stale-disconnect (connId not in liveConns) is a silent no-op", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_2);
+        yield* presence.onAgentConnect(AGENT_A, CONN_2);
         yield* drainMicrotasks;
         h.sent.length = 0;
-        yield* projection.onAgentDisconnect(AGENT_A, CONN_1);
+        yield* presence.onAgentDisconnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("online");
       }),
     );
   });
 });
 
-describe("PresenceProjection — subscriber-snapshot semantics", () => {
+describe("PresenceService — subscriber-snapshot semantics", () => {
   it("only subscribed connections receive `presence/changed`", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_B]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_B]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
       }),
@@ -351,36 +339,34 @@ describe("PresenceProjection — subscriber-snapshot semantics", () => {
   });
 });
 
-// Round-5 codex P2 fix — multi-connection-per-agent correctness.
 // `AgentEndpointResolver` supports an agent holding multiple
-// simultaneous WS connections (web tab + CLI + mobile). The pre-fix
-// projection treated a second simultaneous connect as a "reconnect"
-// — clearing the agent's active leases and emitting a spurious
-// `online` while the agent was still busy on the original conn.
-// These tests pin the new multi-conn semantics down.
-describe("PresenceProjection — multi-connection-per-agent (round-5 codex P2)", () => {
+// simultaneous WS connections (web tab + CLI + mobile). A second
+// simultaneous connect must NOT be treated as a "reconnect" that
+// clears the agent's active leases and emits a spurious `online` while
+// the agent is still busy on the original conn. These tests pin the
+// multi-conn semantics down.
+describe("PresenceService — multi-connection-per-agent", () => {
   it("a second simultaneous WS connection does NOT clobber an active lease (status stays `working`, no emission)", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         // Sanity: status is now `working` (subscriber has seen online
         // + working).
-        const statusOnConn1 = yield* projection.statusOf(AGENT_A);
+        const statusOnConn1 = yield* presence.statusOf(AGENT_A);
         expect(statusOnConn1).toBe("working");
         h.sent.length = 0;
         // Agent opens a second simultaneous connection (e.g. mobile
         // app while the CLI lease is still active).
-        yield* projection.onAgentConnect(AGENT_A, CONN_2);
+        yield* presence.onAgentConnect(AGENT_A, CONN_2);
         yield* drainMicrotasks;
         // The new connection does not affect the agent's lease set.
         // Status stays `working`; the dedup rule elides emission
         // (working → working).
         expect(h.sent).toEqual([]);
-        const statusAfterSecondConnect = yield* projection.statusOf(AGENT_A);
+        const statusAfterSecondConnect = yield* presence.statusOf(AGENT_A);
         expect(statusAfterSecondConnect).toBe("working");
       }),
     );
@@ -388,22 +374,21 @@ describe("PresenceProjection — multi-connection-per-agent (round-5 codex P2)",
 
   it("losing a non-leased simultaneous connection leaves the agent `working` on the other conn (no emission)", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
         // Agent connects on conn1 + conn2; lease granted on conn1
         // only.
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onAgentConnect(AGENT_A, CONN_2);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_2);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         h.sent.length = 0;
         // conn2 disconnects; it had no leases, so no per-conn lease
         // bucket to drop, and conn1's lease is still active.
-        yield* projection.onAgentDisconnect(AGENT_A, CONN_2);
+        yield* presence.onAgentDisconnect(AGENT_A, CONN_2);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("working");
       }),
     );
@@ -411,23 +396,22 @@ describe("PresenceProjection — multi-connection-per-agent (round-5 codex P2)",
 
   it("losing the leased connection while another conn is still live flips status to `online` (lease cleaned up per-conn)", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onAgentConnect(AGENT_A, CONN_2);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_2);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         h.sent.length = 0;
         // conn1 disconnects — its lease bucket is dropped wholesale
         // so the agent's working count drains to zero. conn2 is
         // still live, so the agent stays online (not offline).
-        yield* projection.onAgentDisconnect(AGENT_A, CONN_1);
+        yield* presence.onAgentDisconnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(decodeFrames(h.sent)).toEqual([
           { agentId: AGENT_A, status: "online" },
         ]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("online");
       }),
     );
@@ -435,25 +419,24 @@ describe("PresenceProjection — multi-connection-per-agent (round-5 codex P2)",
 
   it("losing the last live connection flips status to `offline` (entry deleted)", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onAgentConnect(AGENT_A, CONN_2);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_2);
         yield* drainMicrotasks;
         h.sent.length = 0;
         // Disconnect conn1 first — second conn still live, no
         // emission.
-        yield* projection.onAgentDisconnect(AGENT_A, CONN_1);
+        yield* presence.onAgentDisconnect(AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
         // Disconnect conn2 — entry is now empty, emit offline.
-        yield* projection.onAgentDisconnect(AGENT_A, CONN_2);
+        yield* presence.onAgentDisconnect(AGENT_A, CONN_2);
         yield* drainMicrotasks;
         expect(decodeFrames(h.sent)).toEqual([
           { agentId: AGENT_A, status: "offline" },
         ]);
-        const status = yield* projection.statusOf(AGENT_A);
+        const status = yield* presence.statusOf(AGENT_A);
         expect(status).toBe("offline");
       }),
     );
@@ -461,25 +444,24 @@ describe("PresenceProjection — multi-connection-per-agent (round-5 codex P2)",
 
   it("active leases on different simultaneous conns are tracked independently", () => {
     const h = makeHarness();
-    h.subscribers.subscribe(SUBSCRIBER, [AGENT_A]);
-    return withHarness(h, (projection) =>
+    return withHarness(h, [[SUBSCRIBER, [AGENT_A]]], (presence) =>
       Effect.gen(function* () {
-        yield* projection.onAgentConnect(AGENT_A, CONN_1);
-        yield* projection.onAgentConnect(AGENT_A, CONN_2);
-        yield* projection.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
-        yield* projection.onLeaseActiveBegin(LEASE_Y, AGENT_A, CONN_2);
+        yield* presence.onAgentConnect(AGENT_A, CONN_1);
+        yield* presence.onAgentConnect(AGENT_A, CONN_2);
+        yield* presence.onLeaseActiveBegin(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveBegin(LEASE_Y, AGENT_A, CONN_2);
         yield* drainMicrotasks;
         h.sent.length = 0;
         // Releasing LEASE_X on conn1 keeps the agent working (LEASE_Y
         // is still active on conn2).
-        yield* projection.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
+        yield* presence.onLeaseActiveEnd(LEASE_X, AGENT_A, CONN_1);
         yield* drainMicrotasks;
         expect(h.sent).toEqual([]);
-        const stillWorking = yield* projection.statusOf(AGENT_A);
+        const stillWorking = yield* presence.statusOf(AGENT_A);
         expect(stillWorking).toBe("working");
         // Releasing LEASE_Y drains the active set; agent flips to
         // online.
-        yield* projection.onLeaseActiveEnd(LEASE_Y, AGENT_A, CONN_2);
+        yield* presence.onLeaseActiveEnd(LEASE_Y, AGENT_A, CONN_2);
         yield* drainMicrotasks;
         expect(decodeFrames(h.sent)).toEqual([
           { agentId: AGENT_A, status: "online" },
