+++ packages/server/src/transport/layer-tags.ts
@@ -46,10 +46,6 @@ import type {
   SessionValidatorTag,
   AppHostTag,
 } from "../app/layers.js";
-// v7 (codex r6 P2 #2): PresenceProjectionTag joins the network-tier
-// allowlist. Network handlers (presence/subscribe) and the
-// connect-handler at identity-tier yield it.
-import type { PresenceProjectionTag } from "../network/services/presence-projection.js";
 import type {
   TmAuthority,
   TaskReadAccess,
@@ -97,12 +93,7 @@ export type NetworkTags =
   | AgentEndpointResolverTag
   | ConnectionManagerTag
   | NetworkSendServiceTag
-  | PresenceServiceTag
-  // v7 (codex r6 P2 #2): the projection joins the network tier — the
-  // connect-handler yields `PresenceProjectionTag` to call
-  // `onAgentConnect(agentId, connId)`, and the `presence/subscribe`
-  // handler yields it to call `statusMany(visibleIds)`.
-  | PresenceProjectionTag;
+  | PresenceServiceTag;
 
 /**
  * Task-layer allowlist: conversations, messages, tasks.
